+++ src/api/__tests__/open-service-url.test.ts
@@ -145,6 +145,24 @@ describe('api', () => {
       expect(url).toBe(`${VAULT_DOMAINS.production}/customer-support?${query}`);
     });
 
+    test('vault/onboarding', () => {
+      const cobrandClientId = 'cobrandClientId';
+      const sdk = new MtLinkSdk();
+      const locale = 'ja';
+      sdk.init(clientId, { locale, cobrandClientId });
+
+      const url = openServiceUrl(sdk.storedOptions, 'vault', { view: 'onboarding' });
+
+      const query = qs.stringify({
+        client_id: clientId,
+        cobrand_client_id: cobrandClientId,
+        locale,
+        configs: generateConfigs()
+      });
+
+      expect(url).toBe(`${VAULT_DOMAINS.production}/onboarding?${query}`);
+    });
+
     test('link-kit', () => {
       const url = openServiceUrl(new MtLinkSdk().storedOptions, 'link-kit');